Refactor: move download part to another object for dev well-being
