+++ internet/src/main/java/earth/mp3player/internet/UpdateDownloadManager.kt
@@ -1,0 +1,187 @@
+/*
+ * This file is part of MP3 Player.
+ *
+ * MP3 Player is free software: you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License as published by the Free Software Foundation,
+ * either version 3 of the License, or (at your option) any later version.
+ *
+ * MP3 Player is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with MP3 Player.
+ * If not, see <https://www.gnu.org/licenses/>.
+ *
+ * **** INFORMATIONS ABOUT THE AUTHOR *****
+ * The author of this file is Antoine Pirlot, the owner of this project.
+ * You find this original project on github.
+ *
+ * My github link is: https://github.com/antoinepirlot
+ * This current project's link is: https://github.com/antoinepirlot/MP3-Player
+ *
+ * You can contact me via my email: pirlot.antoine@outlook.com
+ * PS: I don't answer quickly.
+ */
+
+package earth.mp3player.internet
+
+import android.app.Activity
+import android.app.DownloadManager
+import android.content.Context
+import android.content.Intent
+import android.content.IntentFilter
+import android.net.Uri
+import android.os.Build
+import android.os.Environment
+import androidx.core.content.getSystemService
+import earth.mp3player.internet.Versions.ALPHA
+import earth.mp3player.internet.Versions.ALPHA_APK_REGEX
+import earth.mp3player.internet.Versions.BETA
+import earth.mp3player.internet.Versions.BETA_APK_REGEX
+import earth.mp3player.internet.Versions.PREVIEW
+import earth.mp3player.internet.Versions.PREVIEW_APK_REGEX
+import earth.mp3player.internet.Versions.RELEASES_URL
+import earth.mp3player.internet.Versions.RELEASE_APK_REGEX
+import earth.mp3player.internet.Versions.versionType
+import earth.mp3player.internet.utils.showToastOnUiThread
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import okhttp3.Response
+
+/**
+ * @author Antoine Pirlot on 14/04/2024
+ */
+object UpdateDownloadManager {
+    private var downloadId: Long = -1
+    private const val MIME_TYPE = "application/vnd.android.package-archive"
+
+    fun downloadUpdateApk(context: Context) {
+        if (UpdateManager.downloadStatus.value == APKDownloadStatus.CHECKING || UpdateManager.downloadStatus.value == APKDownloadStatus.DOWNLOADING) {
+            return
+        }
+        val activity = Activity()
+        CoroutineScope(Dispatchers.IO).launch {
+            UpdateManager.downloadStatus.value = APKDownloadStatus.CHECKING
+            showToastOnUiThread(
+                context = context,
+                activity = activity,
+                message = context.getString(R.string.download_checking)
+            )
+            try {
+                if (UpdateManager.updateAvailableStatus.value != UpdateAvailableStatus.AVAILABLE) {
+                    //Can't be downloaded
+                    UpdateManager.downloadStatus.value = APKDownloadStatus.NOT_FOUND
+                    showToastOnUiThread(
+                        context = context,
+                        activity = activity,
+                        message = context.getString(R.string.download_not_found)
+                    )
+                    return@launch
+                }
+                val downloadUrl: String =
+                    getDownloadUrl(context = context, activity = activity) ?: return@launch
+                val appName: String = downloadUrl.split("/").last().split("_").first()
+                val downloadManager: DownloadManager = context.getSystemService()!!
+                val downloadUri: Uri = Uri.parse(downloadUrl)
+                val req: DownloadManager.Request = DownloadManager.Request(downloadUri)
+                req.setMimeType(MIME_TYPE)
+                val destination =
+                    "file://" + context.getExternalFilesDir(Environment.DIRECTORY_DOWNLOADS)
+                        .toString() + "/$appName"
+                val destinationUri: Uri = Uri.parse(destination)
+                req.setDestinationUri(destinationUri)
+
+                req.setNotificationVisibility(DownloadManager.Request.VISIBILITY_VISIBLE)
+                setDownloadReceiver(context = context)
+                downloadId = downloadManager.enqueue(req)
+                UpdateManager.downloadStatus.value = APKDownloadStatus.DOWNLOADING
+                showToastOnUiThread(
+                    context = context,
+                    activity = activity,
+                    message = context.getString(R.string.downloading)
+                )
+            } catch (_: Exception) {
+                UpdateManager.downloadStatus.value = APKDownloadStatus.FAILED
+                showToastOnUiThread(
+                    context = context,
+                    activity = activity,
+                    message = context.getString(R.string.download_failed)
+                )
+            }
+        }
+    }
+
+    /**
+     * Get the download url for the latest version. It looks like:
+     * "https://github.com/antoinepirlot/MP3-Player/releases/download/vx.y.z-beta/[appName]_vx.y.z[-versionType].apk"
+     *
+     * @return the download url or null if not found
+     */
+    private fun getDownloadUrl(context: Context, activity: Activity): String? {
+        val regex: Regex =
+            when (versionType) {
+                ALPHA -> ALPHA_APK_REGEX
+                BETA -> BETA_APK_REGEX
+                PREVIEW -> PREVIEW_APK_REGEX
+                else -> RELEASE_APK_REGEX
+            }
+        val res: Response = UpdateManager.getUrlResponse(
+            context = context,
+            url = "${RELEASES_URL}/expanded_assets/${UpdateManager.latestVersion.value}"
+        )!!
+        if (!res.isSuccessful) {
+            res.close()
+            UpdateManager.downloadStatus.value = APKDownloadStatus.NOT_FOUND
+            showToastOnUiThread(
+                context = context,
+                activity = activity,
+                message = context.getString(R.string.download_not_found)
+            )
+            return null
+        }
+        val page: String = res.body!!.string()
+        res.close()
+        var apkFileName: String? = regex.find(
+            page,
+            0
+        )?.value
+        if (apkFileName == null) {
+            UpdateManager.downloadStatus.value = APKDownloadStatus.NOT_FOUND
+            showToastOnUiThread(
+                context = context,
+                activity = activity,
+                message = context.getString(R.string.download_not_found)
+            )
+            return null
+        }
+
+        apkFileName = apkFileName.drop(1).dropLast(1) //Avoid > and <
+        return "${RELEASES_URL}/download/${UpdateManager.latestVersion.value}/$apkFileName"
+    }
+
+    //TODO make it available for Android R and later
+    private fun setDownloadReceiver(context: Context) {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            context.registerReceiver(
+                DownloadReceiver, IntentFilter(DownloadManager.ACTION_DOWNLOAD_COMPLETE),
+                Context.RECEIVER_EXPORTED
+            )
+        }
+    }
+
+    /**
+     * Launch the installation procedure by request the user to install the app.
+     */
+    fun installUpdate(context: Context) {
+        val downloadManager: DownloadManager =
+            context.getSystemService(Context.DOWNLOAD_SERVICE) as DownloadManager
+        val contentUri: Uri = downloadManager.getUriForDownloadedFile(downloadId)
+        val install = Intent(Intent.ACTION_VIEW)
+        install.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
+        install.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
+        install.putExtra(Intent.EXTRA_NOT_UNKNOWN_SOURCE, true)
+        install.data = contentUri
+        context.startActivity(install)
+    }
+}
